refactor: added config to separate concerns
